+++ example/lib/main.dart
@@ -5,9 +5,8 @@ main() async {
   runApp(FlutterTeXExample());
 }
 
-TeXViewContainer _customTeXViewChild(String id, String title, String body) {
-  return TeXViewContainer(
-      id: id,
+TeXViewWidget _customTeXViewChild(String title, String body) {
+  return TeXViewColumn(
       style: TeXViewStyle(
         margin: TeXViewMargin.all(10),
         borderRadius: TeXViewBorderRadius.all(25),
@@ -15,14 +14,14 @@ TeXViewContainer _customTeXViewChild(String id, String title, String body) {
         contentColor: Colors.white,
       ),
       children: [
-        TeXViewTeX(title,
+        TeXViewDocument(title,
             style: TeXViewStyle(
               padding: TeXViewPadding.only(top: 5, bottom: 5),
               textAlign: TeXViewTextAlign.Center,
               backgroundColor: Colors.red,
               contentColor: Colors.white,
             )),
-        TeXViewTeX(body,
+        TeXViewDocument(body,
             style: TeXViewStyle.fromCSS(
                 "color:white;background-color:light-green;padding:10px"))
       ]);
@@ -76,6 +75,7 @@ class MainPage extends StatelessWidget {
             TeXExamples.introduction,
             TeXExamples.quadraticEquation,
             TeXExamples.relationEnergyPrincipalQuantum,
+            TeXExamples.alignedTag,
             TeXExamples.bohrRadius,
             TeXExamples.chemistryEquations,
             TeXExamples.matrix,
@@ -89,7 +89,7 @@ class MainPage extends StatelessWidget {
             TeXExamples.introduction,
             TeXExamples.quadraticEquation,
             TeXExamples.relationEnergyPrincipalQuantum,
-            TeXExamples.alignTag,
+            TeXExamples.alignedTag,
             TeXExamples.bohrRadius,
             TeXExamples.chemistryEquations,
             TeXExamples.matrix,
@@ -106,7 +106,7 @@ class MainPage extends StatelessWidget {
       String asset,
       String text,
       TeXViewRenderingEngine renderingEngine,
-      List<TeXViewContainer> teXViewChildren) {
+      List<TeXViewWidget> teXViewChildren) {
     return Padding(
       padding: const EdgeInsets.all(20.0),
       child: RaisedButton(
@@ -159,22 +159,20 @@ class TestExample extends StatelessWidget {
         body: TeXView(
             renderingEngine: TeXViewRenderingEngine.katex(),
             children: [
-              TeXViewContainer(id: "container_0", children: [
-                TeXViewTeX(
+              TeXViewColumn(children: [
+                TeXViewDocument(
                   r"<h3>Quadratic Equation</h3>",
                   style: TeXViewStyle(
                     textAlign: TeXViewTextAlign.Center,
                     backgroundColor: Colors.red,
                     contentColor: Colors.white,
                   ),
-                  id: "tex_0",
                 ),
-                TeXViewTeX(r"""<p>                                
+                TeXViewDocument(r"""<p>                                
                            When \(a \ne 0 \), there are two solutions to \(ax^2 + bx + c = 0\) and they are
                            $$x = {-b \pm \sqrt{b^2-4ac} \over 2a}.$$</p>""",
                     style: TeXViewStyle.fromCSS(
-                        "color:white;background-color:green;padding:20px"),
-                    id: "tex_1"),
+                        "color:white;background-color:green;padding:20px")),
               ])
             ],
             style: TeXViewStyle(
@@ -209,7 +207,7 @@ class TestExample extends StatelessWidget {
 }
 
 class TeXExamples {
-  static TeXViewContainer introduction = TeXViewContainer(
+  static TeXViewWidget introduction = TeXViewColumn(
       style: TeXViewStyle(
           margin: TeXViewMargin.all(10),
           padding: TeXViewPadding.all(10),
@@ -219,38 +217,35 @@ class TeXExamples {
               borderStyle: TeXViewBorderStyle.Solid,
               borderColor: Colors.green))),
       children: [
-        TeXViewTeX(r"""<h2>Flutter \( \rm\\TeX \)</h2>""",
+        TeXViewDocument(r"""<h2>Flutter \( \rm\\TeX \)</h2>""",
             style: TeXViewStyle(
                 borderRadius: TeXViewBorderRadius.only(
                     bottomRight: 20, bottomLeft: 20, topLeft: 10, topRight: 10),
                 textAlign: TeXViewTextAlign.Center,
                 width: 200,
                 margin: TeXViewMargin.zeroAuto(),
-                backgroundColor: Colors.green),
-            id: "title_0"),
-        TeXViewTeX(r"""      
+                backgroundColor: Colors.green)),
+        TeXViewDocument(r"""   
       <p>Flutter \( \rm\\TeX \) is a Flutter Package to render so many types of equations based on \( \rm\\LaTeX \), It also includes full HTML with JavaScript
       support.</p>
-      """, id: "body_0")
+      """)
       ]);
 
-  static TeXViewContainer quadraticEquation =
-      _customTeXViewChild("child_1", r"<h3>Quadratic Equation</h3>", r"""
+  static TeXViewWidget quadraticEquation =
+      _customTeXViewChild(r"<h3>Quadratic Equation</h3>", r"""
      When \(a \ne 0 \), there are two solutions to \(ax^2 + bx + c = 0\) and they are
      $$x = {-b \pm \sqrt{b^2-4ac} \over 2a}.$$<br>""");
 
-  static TeXViewContainer bohrRadius = _customTeXViewChild(
-      "child_2",
+  static TeXViewWidget bohrRadius = _customTeXViewChild(
       r"<h3>Bohr's Radius</h3>",
       r"""\( a_0 = \frac{{\hbar ^2 }}{{m_e ke^2 }} \)""");
 
-  static TeXViewContainer relationEnergyPrincipalQuantum = _customTeXViewChild(
-      "child_3",
+  static TeXViewWidget relationEnergyPrincipalQuantum = _customTeXViewChild(
       r"<h3>Relationship between Energy and Principal Quantum Number</h3>",
       r"""\( E_n = - R_H \left( {\frac{1}{{n^2 }}} \right) = \frac{{ - 2.178 \times 10^{ - 18} }}{{n^2 }}joule \)""");
 
-  static TeXViewContainer chemistryEquations = _customTeXViewChild(
-      "child_4", r"<h3>Chemistry Equations</h3>", r"""<p>    
+  static TeXViewWidget chemistryEquations =
+      _customTeXViewChild(r"<h3>Chemistry Equations</h3>", r"""<p>    
       
      \( \ce{CO2 + C -> 2 CO} \) <br>
      
@@ -260,36 +255,38 @@ class TeXExamples {
      
     </p>""");
 
-  static TeXViewContainer matrix =
-      _customTeXViewChild("child_5", r"<h3>Matrix</h3>", r"""  <p>
+  static TeXViewWidget matrix =
+      _customTeXViewChild(r"<h3>Matrix</h3>", r"""  <p>
      $$ \begin{bmatrix}
          a & b \\
          c & d
       \end{bmatrix}$$</p>""");
 
-  static TeXViewContainer alignTag =
-      _customTeXViewChild("child_6", r"<h3>Align Tag</h3>", r"""    
+  static TeXViewWidget alignedTag =
+      _customTeXViewChild(r"<h3>Aligned Tag</h3>", r"""    
            $$
-           \begin{align}
+           \begin{aligned}
            \dot{x} & = \sigma(y-x) \\
            \dot{y} & = \rho x - y - xz \\
            \dot{z} & = -\beta z + xy
-           \end{align}
-           $$
-    </p>""");
+           \end{aligned}
+           $$""");
 
-  static TeXViewContainer others =
-      _customTeXViewChild("child_7", r"<h3> Others </h3>", r"""<p>    
+  static TeXViewWidget others =
+      _customTeXViewChild(r"<h3> Others </h3>", r"""<p>    
     $$ \oint_C {E \cdot d\ell = - \frac{d}{{dt}}} \int_S {B_n dA} $$<br>
-     $$ y = \frac{{n!}}{{k!\left( {n - k} \right)!}}p^k q^{n - k}  = \left( {\begin{array}{*{20}c}
+    
+     $$
+     y = \frac{{n!}}{{k!\left( {n - k} \right)!}}p^k q^{n - k}  = \left( {\begin{array}{*{20}c}
               n  \\
               k  \\
-            \end{array}} \right)p^k q^{n - k} $$
+            \end{array}} \right)p^k q^{n - k} 
+     $$
     </p>""");
 }
 
 class TeXViewPage extends StatefulWidget {
-  final List<TeXViewContainer> teXViewChildren;
+  final List<TeXViewWidget> teXViewChildren;
   final String label;
   final TeXViewRenderingEngine renderingEngine;
 
@@ -308,6 +305,7 @@ class _TeXViewPageState extends State<TeXViewPage> {
         title: Text("Rendering with ${widget.label} Example"),
       ),
       body: ListView(
+        physics: ScrollPhysics(),
         children: <Widget>[
           Text(
             "Rendered TeX HTML with ${widget.label}",
@@ -319,26 +317,11 @@ class _TeXViewPageState extends State<TeXViewPage> {
               decoration: TextDecoration.underline,
             ),
           ),
-          TeXView(
-            renderingEngine: TeXViewRenderingEngine.katex(),
-            children: [
-              TeXViewContainer(id: "container_0", children: [
-                TeXViewTeX(
-                  '<span>There are 150 objective multiple choice questions in this paper. This Question Booklet is made up of the following five sections : Section-I, 11, 111, IV &amp; V of 30-30 questions each. The candidate has to attempt all sections. Each question is of one mark. There is no negative marking for wrong answer</span>',
-                  style: TeXViewStyle(
-                    textAlign: TeXViewTextAlign.Center,
-                    // backgroundColor: Colors.red,
-                    contentColor: Colors.black,
-                  ),
-                  id: "tex_0",
-                )
-              ])
-            ],
-          ),
+          teXView(),
           TeXView(
               showLoadingWidget: false,
               renderingEngine: widget.renderingEngine,
-              children: widget.teXViewChildren,
+              children: [...widget.teXViewChildren],
               style: TeXViewStyle(
                 margin: TeXViewMargin.all(10),
                 elevation: 10,
@@ -366,5 +349,43 @@ class _TeXViewPageState extends State<TeXViewPage> {
         ],
       ),
     );
+  }
+
+  TeXView teXView() {
+    return TeXView(
+        showLoadingWidget: false,
+        renderingEngine: widget.renderingEngine,
+        children: [
+          TeXViewInkWell(
+            id: "id_0",
+            child: TeXViewImage.network(
+                'https://img.wallpapersafari.com/desktop/1920/1080/84/27/nMWzIB.jpg'),
+            style: TeXViewStyle(borderRadius: TeXViewBorderRadius.all(20)),
+
+          )
+        ],
+        style: TeXViewStyle(
+          elevation: 10,
+          borderRadius: TeXViewBorderRadius.all(25),
+          border: TeXViewBorder.all(TeXViewBorderDecoration(
+              borderColor: Colors.blue,
+              borderStyle: TeXViewBorderStyle.Solid,
+              borderWidth: 5)),
+          backgroundColor: Colors.white,
+        ),
+        loadingWidget: Center(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              CircularProgressIndicator(),
+              Text("Rendering...")
+            ],
+          ),
+        ),
+        onTap: (childID) {
+          print("TeXView $childID is tapped.");
+        });
   }
 }