+++ lib/flutter_tex.dart
@@ -1,7 +1,11 @@
 library flutter_tex;
 
+export 'src/models/tex_view_column.dart';
 export 'src/models/tex_view_container.dart';
-export 'src/models/tex_view_tex.dart';
+export 'src/models/tex_view_document.dart';
+export 'src/models/tex_view_image.dart';
+export 'src/models/tex_view_ink_well.dart';
+export 'src/models/tex_view_widget.dart';
 export 'src/utils/tex_view_rendering_engine.dart';
 export 'src/utils/tex_view_style.dart';
 export 'src/views/platform_interface.dart';