+++ lib/src/models/tex_view_widget.dart
@@ -1,0 +1,3 @@
+abstract class TeXViewWidget {
+  Map<dynamic, dynamic> toJson();
+}